// src/app/models/common/travels_management/advances/advances.ts
export interface Advances {
    curreny_code: string;
    curreny_name: string;
    curreny_symbol: string;
    date_advance: string;
    date_created_at: string;
    id: number;
    level: number;
    observation: string;
    status: string;
    status_name: string;
    travel_request_id: number;
    travel_request_name: string;
    value: number;
}